requirements.txt
streamlit==1.32.0
numpy==1.26.3
pandas==2.2.0
Pillow==10.2.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
scikit-learn==1.4.0
matplotlib==3.8.2 